+++ src/execution/executor_update.h
@@ -112,7 +112,7 @@ class UpdateExecutor : public AbstractExecutor {
         return nullptr;
     }
 
-    size_t tupleLen() const override { throw UnreachableCodeError(); }
+    int tupleLen() const override { throw UnreachableCodeError(); }
 
     const std::vector<ColMeta> &cols() const override { throw UnreachableCodeError(); }
